refactor(character): split CharacterDetailScreen into Header, SheetSections and EditDialogs
